Abstract Types

FileExplorer.ExpNode — Type

ExpNode

Abstract type for representing all the objects in the file system.

FileExplorer.AbstractFolder — Type

AbstractFolder <: ExpNode

Abstract type for representing folder types in the file system.

FileExplorer.AbstractFile — Type

AbstractFile <: ExpNode

Abstract type for representing file types in the file system.

Folder

FileExplorer.Folder — Type

Folder <: AbstractFolder

Represent a folder and its contents as a nested structure.

Properties

Name	Type	Description

path	String	The path to the folder

stat	FS.StatStruct	The stat information

style	Dict{Symbol, AbstractString}	Styling parameters

children	OrderedDict{String, ExpNode}	The contents of the folder

nFolders	Int64	The number of subfolders

nFiles	Int64	The number of files

FileExplorer.Folder — Method

Folder(path::String; hidden=false, lazy=false)

Represent a folder and its contents as a nested structure.

path: The path to the folder.

hidden=false: Whether to include hidden files and folders.

lazy=false: Whether to map contents of the subfolders.

index_files=true: Whether to include files. Set to false to see only the folder structure.

File

FileExplorer.File — Type

File <: AbstractFile

Represent a file in the file system.

Properties

Name	Type	Description

path	String	The path to the file

stat	FS.StatStruct	The stat information

style	Dict{Symbol, AbstractString}	Styling parameters

FileExplorer.File — Method

File(path::String)

Represent a file in the file system.

path: The path to the file.

List elements

FileExplorer.ls — Method

ls(folder::AbstractFolder; kwargs...)

List the contents of a folder in a grid format. Function mimics the behavior of ls command in Unix-like systems. By default, it tries to fit all elements in the most vertically compact layout (with some default settings, like padding size). There is a number of keyword arguments that can be used to customize the output. Most notably, passing long=true will display the content in a single column with typical details like permissions, size, and modification date, similarly to ls -l. The actual displayed information will depend on the operating system.

Possible keyword arguments:

long::Bool: Display the content in a single column view. Default is false.

padding::Int: The number of spaces between the columns. Default is 2.

dims::Int: The direction of the layout. For elements ordered column-wise use 1, for row-wise use 2. Default is 1.

maxWidth::Int: The maximum width of the columns in characters. Default is 30.

show_files::Bool: Whether to show files. Default is true.

unit::Symbol: The unit to use for file sizes. Possible values are :none, :mem, and :memi. Default is :none.

precision::Int: The number of decimal places to show for file sizes. Default is 2.

uid::Bool: Display the user ID. Default is false.

gid::Bool: Display the group ID. Default is false.

header::Bool: Display the header with column names. Default is false.

FileExplorer.dir — Method

dir(path::String; kwargs...)
dir(folder::AbstractFolder; kwargs...)

List the contents of a directory in a long version, similar to the behavior of dir in Windows. Under the hood it is just calling ls with long=true. Other keyword arguments are passed to the ls call. Please look there for more detailed list.

Browse elements

FileExplorer.browse — Method

browse(folderpath::String; hidden=false, lazy=false, kwargs...)

Display the contents of a folder in a tree view. Calling browse directly on a folder path will create a Folder object and pass it to the proper method. See browse(folder::AbstractFolder; kwargs...) for more details on the possible keyword arguments.

path: The path to the folder.

hidden=false: Whether to include hidden files and folders.

lazy=false: Whether to map contents of the subfolders.

show_files=true: Whether to show files.

FileExplorer.browse — Method

browse(folder::AbstractFolder; maxdepth=1, indicate_truncation=false, prefix=" ", kwargs...)

Display the contents of a folder in a tree view. Under the hood, this function is just a thin wrapper around AbstractTrees.print_tree. We provide a default configuration through the keyword arguments mentioned explicitly in the signature. See documentation of AbstractTrees for more details on other possible keyword arguments.

folder: Folder object to display.

maxdepth=1: Maximum depth to display. Defaults to only the immediate children.

indicate_truncation=false: Whether to indicate that subfolders contain more objects.

prefix=" ": Prefix to use for spacing from the tree structure glyphs.

show_files=true: Whether to show files.

Styling

FileExplorer.style — Constant

style::Dict{Symbol, AbstractString}

Dictionary containing styling options implemented in the package. Currently, the following options are available:

Key	Value

:folder_icon	"📁"

:file_icon	"📄"

:folder_color	"blue"

:file_color	"white"